Add Publisher and Connection/Aggregation structs
* Most of iteration 2. This will currently aggregate messages and can be
tested with the commands in `start-it-up.sh`
* Still needs testing, and I'm not sure about these data structures.
They're currently simple at the expense of needing to pass many things
around (e.g. `%Connection{}` and `aggregation.subscribe_routing_keys`
and the like). I imagine this will be useful in the future when there
could be several aggregations with the same connection.
* Need to figure out testing as well. Developing with `start-it-up.sh`
is fine, but I don't want to actually run RabbitMQ as part of the test
suite.